Move under v2_0.common
