+++ instrumentation/ktor/ktor-2-common/library/src/main/kotlin/io/opentelemetry/instrumentation/ktor/v2_0/common/client/AbstractKtorClientTelemetryBuilder.kt
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.instrumentation.ktor.client
+package io.opentelemetry.instrumentation.ktor.v2_0.common.client
 
 import io.ktor.client.request.*
 import io.ktor.client.statement.*
@@ -13,15 +13,14 @@ import io.opentelemetry.api.common.AttributesBuilder
 import io.opentelemetry.context.Context
 import io.opentelemetry.instrumentation.api.incubator.builder.internal.DefaultHttpClientInstrumenterBuilder
 import io.opentelemetry.instrumentation.api.instrumenter.AttributesExtractor
-import io.opentelemetry.instrumentation.ktor.internal.KtorBuilderUtilOld
+import io.opentelemetry.instrumentation.ktor.v2_0.common.internal.KtorBuilderUtil
 
-@Deprecated("Use AbstractKtorClientTelemetryBuilder instead", ReplaceWith("AbstractKtorClientTelemetryBuilder"))
-abstract class AbstractKtorClientTracingBuilder(
+abstract class AbstractKtorClientTelemetryBuilder(
   private val instrumentationName: String
 ) {
   companion object {
     init {
-      KtorBuilderUtilOld.clientBuilderExtractor = { it.clientBuilder }
+      KtorBuilderUtil.clientBuilderExtractor = { it.clientBuilder }
     }
   }
 
@@ -41,47 +40,17 @@ abstract class AbstractKtorClientTracingBuilder(
     return openTelemetry
   }
 
-  @Deprecated(
-    "Please use method `capturedRequestHeaders`",
-    ReplaceWith("capturedRequestHeaders(headers.asIterable())")
-  )
-  fun setCapturedRequestHeaders(vararg headers: String) = capturedRequestHeaders(headers.asIterable())
-
-  @Deprecated(
-    "Please use method `capturedRequestHeaders`",
-    ReplaceWith("capturedRequestHeaders(headers)")
-  )
-  fun setCapturedRequestHeaders(headers: List<String>) = capturedRequestHeaders(headers)
-
   fun capturedRequestHeaders(vararg headers: String) = capturedRequestHeaders(headers.asIterable())
 
   fun capturedRequestHeaders(headers: Iterable<String>) {
     clientBuilder.setCapturedRequestHeaders(headers.toList())
   }
 
-  @Deprecated(
-    "Please use method `capturedResponseHeaders`",
-    ReplaceWith("capturedResponseHeaders(headers.asIterable())")
-  )
-  fun setCapturedResponseHeaders(vararg headers: String) = capturedResponseHeaders(headers.asIterable())
-
-  @Deprecated(
-    "Please use method `capturedResponseHeaders`",
-    ReplaceWith("capturedResponseHeaders(headers)")
-  )
-  fun setCapturedResponseHeaders(headers: List<String>) = capturedResponseHeaders(headers)
-
   fun capturedResponseHeaders(vararg headers: String) = capturedResponseHeaders(headers.asIterable())
 
   fun capturedResponseHeaders(headers: Iterable<String>) {
     clientBuilder.setCapturedResponseHeaders(headers.toList())
   }
-
-  @Deprecated(
-    "Please use method `knownMethods`",
-    ReplaceWith("knownMethods(knownMethods)")
-  )
-  fun setKnownMethods(knownMethods: Set<String>) = knownMethods(knownMethods)
 
   fun knownMethods(vararg methods: String) = knownMethods(methods.asIterable())
 
@@ -92,19 +61,6 @@ abstract class AbstractKtorClientTracingBuilder(
 
   fun knownMethods(methods: Iterable<String>) {
     clientBuilder.setKnownMethods(methods.toSet())
-  }
-
-  @Deprecated("Please use method `attributeExtractor`")
-  fun addAttributesExtractors(vararg extractors: AttributesExtractor<in HttpRequestData, in HttpResponse>) = addAttributesExtractors(extractors.asList())
-
-  @Deprecated("Please use method `attributeExtractor`")
-  fun addAttributesExtractors(extractors: Iterable<AttributesExtractor<in HttpRequestData, in HttpResponse>>) {
-    extractors.forEach {
-      attributeExtractor {
-        onStart { it.onStart(attributes, parentContext, request) }
-        onEnd { it.onEnd(attributes, parentContext, request, response, error) }
-      }
-    }
   }
 
   fun attributeExtractor(extractorBuilder: ExtractorBuilder.() -> Unit = {}) {
@@ -156,19 +112,10 @@ abstract class AbstractKtorClientTracingBuilder(
   )
 
   /**
-   * Configures the instrumentation to emit experimental HTTP client metrics.
-   *
-   * @param emitExperimentalHttpClientMetrics `true` if the experimental HTTP client metrics are to be emitted.
+   * Can be used via the unstable method {@link
+   * Experimental#setEmitExperimentalHttpClientMetrics(AbstractKtorClientTelemetryBuilder, boolean)}.
    */
-  @Deprecated("Please use method `Experimental.emitExperimentalHttpClientMetrics`")
-  fun setEmitExperimentalHttpClientMetrics(emitExperimentalHttpClientMetrics: Boolean) {
-    if (emitExperimentalHttpClientMetrics) {
-      emitExperimentalHttpClientMetrics()
-    }
-  }
-
-  @Deprecated("Please use method `Experimental.emitExperimentalHttpClientMetrics`")
-  fun emitExperimentalHttpClientMetrics() {
+  internal fun emitExperimentalHttpClientMetrics() {
     clientBuilder.setEmitExperimentalHttpClientMetrics(true)
   }
 }
