+++ instrumentation/ktor/ktor-2-common/library/src/main/kotlin/io/opentelemetry/instrumentation/ktor/v2_0/common/client/KtorHttpClientAttributesGetter.kt
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.instrumentation.ktor.client
+package io.opentelemetry.instrumentation.ktor.v2_0.common.client
 
 import io.ktor.client.request.*
 import io.ktor.client.statement.*
