+++ instrumentation/ktor/ktor-2-common/library/src/main/kotlin/io/opentelemetry/instrumentation/ktor/v2_0/common/internal/KtorBuilderUtilOld.kt
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.instrumentation.ktor.internal
+package io.opentelemetry.instrumentation.ktor.v2_0.common.internal
 
 import io.ktor.client.request.*
 import io.ktor.client.statement.*
@@ -11,16 +11,17 @@ import io.ktor.server.request.*
 import io.ktor.server.response.*
 import io.opentelemetry.instrumentation.api.incubator.builder.internal.DefaultHttpClientInstrumenterBuilder
 import io.opentelemetry.instrumentation.api.incubator.builder.internal.DefaultHttpServerInstrumenterBuilder
-import io.opentelemetry.instrumentation.ktor.client.AbstractKtorClientTelemetryBuilder
-import io.opentelemetry.instrumentation.ktor.server.AbstractKtorServerTelemetryBuilder
+import io.opentelemetry.instrumentation.ktor.v2_0.common.client.AbstractKtorClientTracingBuilder
+import io.opentelemetry.instrumentation.ktor.v2_0.common.server.AbstractKtorServerTracingBuilder
 
 /**
  * This class is internal and is hence not for public use. Its APIs are unstable and can change at
  * any time.
  */
-object KtorBuilderUtil {
-  lateinit var clientBuilderExtractor: (AbstractKtorClientTelemetryBuilder) -> DefaultHttpClientInstrumenterBuilder<HttpRequestData, HttpResponse>
+@Deprecated("Use KtorBuilderUtil instead", ReplaceWith("KtorBuilderUtil"))
+object KtorBuilderUtilOld {
+  lateinit var clientBuilderExtractor: (AbstractKtorClientTracingBuilder) -> DefaultHttpClientInstrumenterBuilder<HttpRequestData, HttpResponse>
   lateinit var serverBuilderExtractor: (
-    AbstractKtorServerTelemetryBuilder
+    AbstractKtorServerTracingBuilder
   ) -> DefaultHttpServerInstrumenterBuilder<ApplicationRequest, ApplicationResponse>
 }
