+++ instrumentation/ktor/ktor-2-common/library/src/main/kotlin/io/opentelemetry/instrumentation/ktor/v2_0/common/internal/KtorClientTracingUtil.kt
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.instrumentation.ktor.internal
+package io.opentelemetry.instrumentation.ktor.v2_0.common.internal
 
 import io.ktor.client.*
 import io.ktor.client.request.*
@@ -13,7 +13,7 @@ import io.ktor.util.pipeline.*
 import io.opentelemetry.context.Context
 import io.opentelemetry.extension.kotlin.asContextElement
 import io.opentelemetry.instrumentation.api.semconv.http.HttpClientRequestResendCount
-import io.opentelemetry.instrumentation.ktor.client.AbstractKtorClientTelemetry
+import io.opentelemetry.instrumentation.ktor.client.AbstractKtorClientTracing
 import kotlinx.coroutines.InternalCoroutinesApi
 import kotlinx.coroutines.job
 import kotlinx.coroutines.launch
@@ -23,15 +23,16 @@ import kotlinx.coroutines.withContext
  * This class is internal and is hence not for public use. Its APIs are unstable and can change at
  * any time.
  */
-object KtorClientTelemetryUtil {
+@Deprecated("Use KtorClientTelemetryUtil instead", ReplaceWith("KtorClientTelemetryUtil"))
+object KtorClientTracingUtil {
   private val openTelemetryContextKey = AttributeKey<Context>("OpenTelemetry")
 
-  fun install(plugin: AbstractKtorClientTelemetry, scope: HttpClient) {
+  fun install(plugin: AbstractKtorClientTracing, scope: HttpClient) {
     installSpanCreation(plugin, scope)
     installSpanEnd(plugin, scope)
   }
 
-  private fun installSpanCreation(plugin: AbstractKtorClientTelemetry, scope: HttpClient) {
+  private fun installSpanCreation(plugin: AbstractKtorClientTracing, scope: HttpClient) {
     val initializeRequestPhase = PipelinePhase("OpenTelemetryInitializeRequest")
     scope.requestPipeline.insertPhaseAfter(HttpRequestPipeline.State, initializeRequestPhase)
 
@@ -64,7 +65,7 @@ object KtorClientTelemetryUtil {
   }
 
   @OptIn(InternalCoroutinesApi::class)
-  private fun installSpanEnd(plugin: AbstractKtorClientTelemetry, scope: HttpClient) {
+  private fun installSpanEnd(plugin: AbstractKtorClientTracing, scope: HttpClient) {
     val endSpanPhase = PipelinePhase("OpenTelemetryEndSpan")
     scope.receivePipeline.insertPhaseBefore(HttpReceivePipeline.State, endSpanPhase)
