+++ instrumentation/ktor/ktor-2-common/library/src/main/kotlin/io/opentelemetry/instrumentation/ktor/v2_0/common/server/KtorHttpServerAttributesGetter.kt
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.instrumentation.ktor.server
+package io.opentelemetry.instrumentation.ktor.v2_0.common.server
 
 import io.ktor.server.plugins.*
 import io.ktor.server.request.*
