+++ instrumentation/ktor/ktor-3.0/library/src/main/kotlin/io/opentelemetry/instrumentation/ktor/v3_0/client/KtorClientTracingBuilder.kt
@@ -5,7 +5,7 @@
 
 package io.opentelemetry.instrumentation.ktor.v3_0.client
 
-import io.opentelemetry.instrumentation.ktor.client.AbstractKtorClientTracingBuilder
+import io.opentelemetry.instrumentation.ktor.v2_0.common.client.AbstractKtorClientTracingBuilder
 import io.opentelemetry.instrumentation.ktor.v3_0.InstrumentationProperties.INSTRUMENTATION_NAME
 
 @Deprecated("Use KtorClientTelemetryBuilder instead", ReplaceWith("KtorClientTelemetryBuilder"))
